refactor: auto-split test_pose6dof_placement.py
